# perf(storage): improve prefix deletion in Cache