--- a/tsdb/tsm1/engine_delete_prefix.go
+++ b/tsdb/tsm1/engine_delete_prefix.go
@@ -76,6 +76,7 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 		// TODO(edd): tracing this deep down is currently speculative, so I have
 		// not added the tracing into the TSMReader API.
 		span, _ := tracing.StartSpanFromContextWithOperationName(rootCtx, "TSMFile delete prefix")
+		span.LogKV("file_path", r.Path())
 		defer span.Finish()
 
 		return r.DeletePrefix(name, min, max, pred, func(key []byte) {
@@ -87,13 +88,12 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 		return err
 	}
 
+	span, _ = tracing.StartSpanFromContextWithOperationName(rootCtx, "Cache find delete keys")
+	span.LogKV("cache_size", e.Cache.Size())
+	var keysChecked int // For tracing information.
 	// ApplySerialEntryFn cannot return an error in this invocation.
 	_ = e.Cache.ApplyEntryFn(func(k []byte, _ *entry) error {
-		// TODO(edd): tracing this deep down is currently speculative, so I have
-		// not added the tracing into the Cache API.
-		span, _ := tracing.StartSpanFromContextWithOperationName(rootCtx, "Cache find delete keys")
-		defer span.Finish()
-
+		keysChecked++
 		if !bytes.HasPrefix(k, name) {
 			return nil
 		}
@@ -107,11 +107,10 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 
 		return nil
 	})
+	span.LogKV("cache_cardinality", keysChecked)
+	span.Finish()
 
-	sortSpan, _ := tracing.StartSpanFromContextWithOperationName(rootCtx, "Cache sort keys")
-	sortSpan.Finish()
-
-	// Delete from the cache.
+	// Delete from the cache (traced in cache).
 	e.Cache.DeleteBucketRange(ctx, name, min, max, pred)
 
 	// Now that all of the data is purged, we need to find if some keys are fully deleted
@@ -120,11 +119,13 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 		// TODO(edd): tracing this deep down is currently speculative, so I have
 		// not added the tracing into the Engine API.
 		span, _ := tracing.StartSpanFromContextWithOperationName(rootCtx, "TSMFile determine fully deleted")
+		span.LogKV("file_path", r.Path())
 		defer span.Finish()
 
 		possiblyDead.RLock()
 		defer possiblyDead.RUnlock()
 
+		var keysChecked int
 		iter := r.Iterator(name)
 		for i := 0; iter.Next(); i++ {
 			key := iter.Key()
@@ -149,19 +150,18 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 				possiblyDead.RLock()
 			}
 		}
-
+		span.LogKV("keys_checked", keysChecked)
 		return iter.Err()
 	}); err != nil {
 		return err
 	}
 
+	span, _ = tracing.StartSpanFromContextWithOperationName(rootCtx, "Cache find delete keys")
+	span.LogKV("cache_size", e.Cache.Size())
+	keysChecked = 0
 	// ApplySerialEntryFn cannot return an error in this invocation.
 	_ = e.Cache.ApplyEntryFn(func(k []byte, _ *entry) error {
-		// TODO(edd): tracing this deep down is currently speculative, so I have
-		// not added the tracing into the Cache API.
-		span, _ := tracing.StartSpanFromContextWithOperationName(rootCtx, "Cache find delete keys")
-		defer span.Finish()
-
+		keysChecked++
 		if !bytes.HasPrefix(k, name) {
 			return nil
 		}
@@ -172,6 +172,8 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 		delete(possiblyDead.keys, string(k))
 		return nil
 	})
+	span.LogKV("cache_cardinality", keysChecked)
+	span.Finish()
 
 	if len(possiblyDead.keys) > 0 {
 		buf := make([]byte, 1024)
@@ -218,7 +220,8 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 
 			// Iterate over the series ids we previously extracted from the index
 			// and remove from the series file.
-			span, _ = tracing.StartSpanFromContextWithOperationName(rootCtx, "SFile Delete Series ID")
+			span, _ = tracing.StartSpanFromContextWithOperationName(rootCtx, "SFile Delete Series IDs")
+			span.LogKV("series_id_set_size", set.Cardinality())
 			set.ForEachNoLock(func(id tsdb.SeriesID) {
 				if err = e.sfile.DeleteSeriesID(id); err != nil {
 					return
@@ -230,6 +233,7 @@ func (e *Engine) DeletePrefixRange(rootCtx context.Context, name []byte, min, ma
 
 		// This is the slow path, when not dropping the entire bucket (measurement)
 		span, _ = tracing.StartSpanFromContextWithOperationName(rootCtx, "TSI/SFile Delete keys")
+		span.LogKV("keys_to_delete", len(possiblyDead.keys))
 		for key := range possiblyDead.keys {
 			// TODO(jeff): ugh reduce copies here
 			keyb := []byte(key)